requirements.txt
Flask==2.2.5
numpy==2.0.0
pandas==2.2.2
pyngrok==7.1.6
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.2.2
